Create Production-branch.html
The production environment is the live, operating version of the software. Only stable and extensively tested code is deployed to the production environment to reduce the possibility of problems for end users. This ensures that the code in production is consistent and dependable. Additionally, this corresponds to 'master' branch where it represents the production-ready code.